+++ src/peers/index.ts
@@ -1,5 +1,5 @@
-import PeersContainer from './peersContainer'
 import { getPeersData } from './peersActions'
+import PeersContainer from './peersContainer'
 import peers from './peersReducer'
 
 export { getPeersData, peers }
